fix: use model image metadata for previews and add type icon fallbacks
