+++ src/components/modelBrowser/ModelListItem.vue
@@ -27,7 +27,14 @@
         v-if="imageError || !model.previewUrl"
         class="flex size-full cursor-pointer items-center justify-center bg-gradient-to-br from-smoke-400 via-smoke-800 to-charcoal-400"
       >
-        <!-- Fallback can be empty or show icon -->
+        <i
+          :class="
+            cn(
+              getModelTypeIcon(model.type),
+              'size-6 sm:size-10 text-muted-foreground'
+            )
+          "
+        />
       </div>
       <img
         v-else
@@ -105,6 +112,7 @@ import {
   formatFileSize,
   formatModifiedDate
 } from '@/utils/modelBrowser/modelTransform'
+import { getModelTypeIcon } from '@/utils/modelBrowser/modelTypeIcons'
 import { cn } from '@/utils/tailwindUtil'
 
 const { model, focused = false } = defineProps<{
